Get rid of SnapletHeist newtype
The previous implementation did not work for compiled splices and my attempts
at generalizing it to work with both interpreted and compiled splices all
failed.  In some ways this is better anyway because it means we're always
working in HeistT and don't have to deal with the pain of not having a
MonadHeist type class.