Simplify deepseq workaround
Calculating the length of a lazy `ByteString` forces it deeply (each chunk has to be examined and their lengths summed). Thus we can use `BSL.length` instead of copying the `deepseq` definition.

Really, we *should* use `deepseq` when `bytestring` is new enough, in case some future change invalidates this workaround.